safari-600.7-branch/LayoutTests/js/dom/call-base-resolution-expected.txt

PASS ooaoboabuuaubuabooaoboab should be ooaoboabuuaubuabooaoboab, and is.
PASS successfullyParsed is true

TEST COMPLETE
